+++ NEWS.md
@@ -1,5 +1,15 @@
 ClimaAnalysis.jl Release Notes
 ===============================
+v0.5.14
+-------
+
+## Partial resampling
+When resampling, it may be helpful to resampling only over physical dimensions while
+ignoring the rest of the dimensions. There is a new keyword argument `dim_names` for
+`resampled_as`, which allows the user to resample only over the dimensions in
+`resampled_as`. For example, if one wants to resample only over longitude and latitude, then
+one can use `resampled_as(src_var, dest_var, dim_names = ["longitude, "latitude"])`.
+
 v0.5.13
 -------